Add 'interactive-video-rotation-renamer.py'.
Adds a new script, 'interactive-video-rotation-renamer.py'.

Temporary helper for renaming videos prior to rotating.
Uses mplayer to preview the videos. VLC does auto-rotation from metadata,
while mplayer seems to play the file as-is.

After mplayer has finished, the user selects from a menu.
The file is renamed to reflect the user selection.

A proper solution to this would be to do both the preview, selection and
actual processing at once. But this is at least better than an all manual
approach for now.